+++ packages/dialtone-vue/recipes/conversation_view/editor/editor.vue
@@ -177,6 +177,81 @@
         <div class="d-recipe-editor__button-group-divider" />
       </dt-stack>
       <dt-stack
+        v-if="variableButton.showBtn"
+        direction="row"
+        gap="300"
+      >
+        <dt-popover
+          padding="small"
+          navigation-type="arrow-keys"
+          :modal="false"
+          placement="bottom-start"
+        >
+          <template #anchor="{ attrs }">
+            <dt-tooltip
+              :message="variableButton.tooltipMessage"
+              placement="top"
+            >
+              <template #anchor>
+                <dt-button
+                  v-bind="attrs"
+                  kind="muted"
+                  size="xs"
+                  importance="clear"
+                  :aria-label="variableButton.tooltipMessage"
+                  :data-qa="variableButton.dataQA"
+                  label-class="d-jc-flex-start"
+                >
+                  <template #icon>
+                    <component
+                      :is="variableButton.icon"
+                      size="200"
+                    />
+                  </template>
+                </dt-button>
+              </template>
+            </dt-tooltip>
+          </template>
+          <template #content="{ close }">
+            <dt-input
+              v-model="variableSearchValue"
+              root-class="d-p8 d-pb4 d-w264"
+              type="search"
+              :placeholder="i18n.$t('DIALTONE_EDITOR_VARIABLE_POPOVER_SEARCH_PLACEHOLDER')"
+              size="md"
+              role="menuitem"
+            >
+              <template #leftIcon="{ iconSize }">
+                <dt-icon-search :size="iconSize" />
+              </template>
+            </dt-input>
+            <dt-list-item-group
+              v-for="(category, index) in filteredCategories"
+              :key="category.name"
+              :heading="category.name"
+              heading-class="d-headline--sm-compact d-p8"
+            >
+              <dt-list-item
+                v-for="item in getFilteredItemsForCategory(category)"
+                :key="category.name + item.name"
+                role="menuitem"
+                navigation-type="arrow-keys"
+                @click="
+                  insertVariable(category.name, item);
+                  close();
+                "
+              >
+                {{ item.name }}
+              </dt-list-item>
+              <dt-dropdown-separator
+                v-if="index < filteredCategories.length - 1"
+              />
+            </dt-list-item-group>
+          </template>
+        </dt-popover>
+        <div class="d-recipe-editor__button-group-divider" />
+      </dt-stack>
+      <dt-stack
         v-if="linkButton.showBtn"
         direction="row"
         gap="300"
@@ -292,6 +367,8 @@
         :allow-font-family="true"
         :allow-inline-images="true"
         :allow-line-breaks="true"
+        :allow-variable="true"
+        :variable-items="flattenedVariableItems"
         :hide-link-bubble-menu="true"
         :auto-focus="autoFocus"
         :editable="editable"
@@ -330,6 +407,8 @@ import { DtStack } from '@/components/stack';
 import { DtInput } from '@/components/input';
 import { DtTooltip } from '@/components/tooltip';
 import { DtListItem } from '@/components/list_item';
+import {DtDropdownSeparator} from '@/components/dropdown/index.js';
+import {DtListItemGroup} from '@/components/list_item_group/index.js';
 import {
   DtIconAlignCenter,
   DtIconAlignJustify,
@@ -347,6 +426,7 @@ import {
   DtIconStrikethrough,
   DtIconUnderline,
   DtIconType,
+  DtIconBraces,
   DtIconSearch,
 } from '@dialpad/dialtone-icons/vue3';
 import { DialtoneLocalization } from '@/localization';
@@ -356,6 +436,9 @@ export default {
   name: 'DtRecipeEditor',
 
   components: {
+    DtListItemGroup,
+    DtDropdownSeparator,
+    DtListItem,
     DtRichTextEditor,
     DtButton,
     DtPopover,
@@ -380,6 +463,7 @@ export default {
     DtIconImage,
     DtIconSearch,
     DtIconType,
+    DtIconBraces,
   },
 
   mixins: [],
@@ -583,6 +667,22 @@ export default {
     },
 
     /**
+     * Show button to add a variable
+     */
+    showVariableButton: {
+      type: Boolean,
+      default: false,
+    },
+
+    /**
+     * Variable categories to display when variable button is clicked
+     */
+    variableCategories: {
+      type: Array,
+      default: () => [],
+    },
+
+    /**
      * Show add link default config.
      */
     showAddLink: {
@@ -721,6 +821,7 @@ export default {
       fontStyleSearch: '',
       linkInput: '',
       currentButtonRefIndex: 0,
+      variableSearchValue: '',
       i18n: new DialtoneLocalization(),
     };
   },
@@ -732,6 +833,13 @@ export default {
 
     htmlOutputFormat () {
       return RICH_TEXT_EDITOR_OUTPUT_FORMATS[2];
+    },
+
+    flattenedVariableItems () {
+      if (!this.variableCategories) return [];
+      return this.variableCategories.reduce((acc, category) => {
+        return acc.concat(category.items || []);
+      }, []);
     },
 
     showingTextFormatButtons () {
@@ -936,6 +1044,17 @@ export default {
       };
     },
 
+    variableButton() {
+      return {
+        showBtn: this.showVariableButton,
+        selector: 'variable',
+        icon: DtIconBraces,
+        dataQA: 'dt-recipe-editor-variable-btn',
+        tooltipMessage: this.i18n.$t('DIALTONE_EDITOR_VARIABLE_BUTTON_LABEL'),
+      }
+    },
+
+
     confirmSetLinkButtonLabels () {
       return this.i18n.$ta('DIALTONE_EDITOR_CONFIRM_SET_LINK_BUTTON');
     },
@@ -956,6 +1075,11 @@ export default {
       const searchValue = this.fontStyleSearch.toLowerCase();
       return this.fontStyles.filter((item) =>
         item.name.toLowerCase().includes(searchValue),
+      );
+    },
+    filteredCategories() {
+      return this.variableCategories.filter(
+        (category) => this.getFilteredItemsForCategory(category).length,
       );
     },
   },
@@ -1087,6 +1211,15 @@ export default {
       this.$emit('inline-image-click');
     },
 
+    insertVariable (categoryName, variableData) {
+      // Insert a variable using the custom command from the Variable extension
+      this.$refs.richTextEditor?.editor.chain().focus().insertVariable({
+        id: variableData.id,
+        placeholder: variableData.placeholder || '',
+        altText: '',
+      }).run();
+    },
+
     insertInlineImage (imageUrl) {
       this.$refs.richTextEditor?.editor.chain().focus().setImage({ src: imageUrl }).run();
     },
@@ -1186,6 +1319,16 @@ export default {
         return !this.$refs.richTextEditor?.editor?.getAttributes('textStyle')?.fontSize;
       }
       return this.$refs.richTextEditor?.editor?.isActive('textStyle', { fontSize });
+    },
+    
+    getFilteredItemsForCategory(category) {
+      const searchValue = this.variableSearchValue.toLowerCase();
+      if (category.name.toLowerCase().includes(searchValue)) {
+        return category.items;
+      }
+      return category.items.filter((item) =>
+        item.name.toLowerCase().includes(searchValue),
+      );
     },
   },
 };